Moves resources around, simplifies code. Refs #42.
